backend/pyproject.toml
[project]
name = "smart-english-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "PyJWT>=2.10",
    "passlib[bcrypt]>=1.7",
    "redis>=5.0",
    "sse-starlette>=2.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
